feat(core): Funcion helper para lanzar servidor
